+++ src/patterns/stylesheets/test/include.spec.js
@@ -1,8 +1,8 @@
 require("sass");
 const path = require("path");
-const { runGulp, distScssPath, render } = require("./util");
+const { runGulp, distScssPath, render } = require("../../utils/test/util");
 
-const includePath = path.resolve(path.join(__dirname, "../../src/stylesheets"));
+const includePath = path.resolve(path.join(__dirname, "../"));
 
 describe("include paths", () => {
   it('can be loaded with @import "uswds"', async () => {
