+++ src/patterns/utils/test/util.js
@@ -2,7 +2,7 @@ const path = require("path");
 const child = require("child_process");
 const sass = require("sass");
 
-exports.distPath = path.resolve(path.join(__dirname, "../../dist"));
+exports.distPath = path.resolve(path.join(__dirname, "../../../../dist"));
 exports.distCssPath = path.join(exports.distPath, "css");
 exports.distScssPath = path.join(exports.distPath, "scss");
 exports.runGulp = task =>
